requirements.txt
# API
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# Database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
pgvector==0.2.5

# Neo4j
neo4j==5.19.0

# Task queue
celery==5.4.0
redis==5.0.4

# AI / Embeddings
openai==1.30.1          # OpenRouter uses OpenAI-compatible API
sentence-transformers==3.0.1
torch==2.3.0
spacy==3.7.4

# Document parsing
pymupdf==1.24.3          # PDFs
python-docx==1.1.0       # DOCX
beautifulsoup4==4.12.3   # Web
httpx==0.27.0            # Async HTTP
markdownify==0.12.1

# Graph algorithms
networkx==3.3
python-louvain==0.16

# Utilities
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
tenacity==8.3.0          # Retry logic
structlog==24.1.0        # Structured logging
rich==13.7.1

# Testing
pytest==8.2.0
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0            # for async test client
